Fix evil-quit for clients with multiple frames.
Initially, if one calls evil-quit and the current frame belongs to some
client (if `selected-frame' has the `frame-parameter' 'client), it
closes the client with a call to evil-quitall . If make-frame was called
from a client, both frames share the same proc id, so both are
killed. This fix makes evil-quit call evil-quitall only if the current
frame is the only one that belongs to some client.